Add simulated analytics DB for BQ DML samples.
This generates a random MySQL database (tables defined in
create_sample_db.sql) with random user actions. The reason for putting
these into a SQL database instead of directly into BigQuery is that it
will be used to show export form a SQL database into BigQuery.
Hopefully the limits on query sizes for BigQuery are large enough that
this works for larger databases.

Change-Id: I446c3af72dab60d9ed79a2c814a68f05801ae17b