Add BigQuery DML sample for inserts.
This sample reads a SQL file (for example: one that was output from
mysqldump) and executes each line as a query. At least in my
configuration of mysqldump, each insert statement was on a single line,
so I was able to write data from MySQL to BigQuery with this sample.

Change-Id: Id14b648b0ce6bac651e436d402f480c56d80bd37