Composer: add simple workflow example. Use 1.10beta instead of git. (#1469)
The `composer_simple` sample is intended to demonstrate the concepts of
constructing Apache Airflow DAGs rather than any specific set of
operators.

By using the 1.10beta instead of git for the dependency, we have a
stabler package than building on master. Also, this will be easier for
Travis to cache.