Changed region tags in sample
Currently none of the Endpoints Frameworks docs are using the "echo_api" region tag in this sample. I would like to show just a small snippet of code, so I added the "echo_api_method" region tag. In case we want to show the entire class in the docs, I've renamed the "echo_api" region tag to "echo_api_class".