Create an example of how to use the OS Login API to manage SSH keys for a service account.
In some situations you might want a Compute Engine service account to be able to SSH to instances over an internal VPC network to run automated tasks. If your project uses OS Login, your service account can generate a key pair for itself and apply the public key as an authorized key on its account. With OS Login enabled on a project or an instance, the instance can access the public key, grant SSH access to the service account, and allow the service account to run remote commands over SSH.