[firestore] fix: mark 3 tests as flaky
fixes #3131

These tests are testing sample code that has `sleep(1)` in it.
The test is naturally flaky, but when I run the test 100 times, they all
pass, so they are somewhat stable.

I believe it will mitigate the flakiness a lot just by retrying 3 times.

If it doesn't work, we'll probably need to make a fundamental fix to
remove `sleep(1)` calls.