remove misleading raise_for_status (#2948)
When the response has error, the error detail is not exposed because of raise_for_status, so remove it and show the detailed message.

For instance, the detailed error is UNAUTHENTICATED:Method doesn't allow unregistered callers (callers without established identity). Please use API Key or other form of API consumer identity to call this API.
but we only got requests.exceptions.HTTPError: 401 Client Error: Unauthorized for url.