test: run asserts involving NDB entities in an NDB context
When an assert fails, pytest will attempt to do magic to provide
meaningful output. When one of the objects in an assert is an NDB
entity, this can lead to a `ContextError` if the assert isn't run in an
NDB context, which creates a red herring, obscuring the actual failure
and sending the developer looking for bugs that aren't there.

See #5681, #5682