fix: 3.10 build - Remove noxfiles, add noxfile_configs where needed (#7077)
Some tests in the build were failing because nox failed to find a session. This is because there were still some noxfiles hanging around where there shouldn't have been anymore. I removed them and in cases where the noxfile specified nox version, added a noxfile config. 

I added 3.8 back to the pubsub/streaming analytics because I know 3.8 is compatible with Dataflow, but the versions ignored in this noxfile config are the ones that are not

If you want to look at 3.10 builds where the logs are truncated see [this tip](https://github.com/GoogleCloudPlatform/python-docs-samples/issues/6902#issuecomment-964677619)

Relates to #6902