update Flask to 2.0.3 for flex codelab demo
App Engine Flex doesn't any Python version greater than 3.7 selectable by default as its runtime. When trying to deploy the app with the current requirements, we get this error:

```
ERROR: Could not find a version that satisfies the requirement Flask==2.1.0 (from -r requirements.txt (line 1)) (from versions: 0.1, 0.2, 0.3, 0.3.1, 0.4, 0.5, 0.5.1,
 0.5.2, 0.6, 0.6.1, 0.7, 0.7.1, 0.7.2, 0.8, 0.8.1, 0.9, 0.10, 0.10.1, 0.11, 0.11.1, 0.12, 0.12.1, 0.12.2, 0.12.3, 0.12.4, 0.12.5, 1.0, 1.0.1, 1.0.2, 1.0.3, 1.0.4, 1.1.0, 1.1.1, 1.1.2, 1.1.3, 
1.1.4, 2.0.0rc1, 2.0.0rc2, 2.0.0, 2.0.1, 2.0.2, 2.0.3)
```

As a result, the training labs that use this demo are broken. Changing to Flask `v2.0.3` fixes the demo.