dataflow (fix): pin in to ubuntu:impish
It looks like on the newly released `ubuntu:jammy` which is what `ubuntu:latest` points to now, the available Python installation is Python 3.10. Even though Apache Beam is available in Python 3.10 through `pip`, there is no "official" support for it so there is no `apache/beam_python3.10_sdk` image.

So we're pinning to `ubuntu:impish` which is the latest ubuntu image with Python 3.9 support.