refactor(compute-oslogin): Rewriting the OSLogin SSH sample and it's test (#8444)
## Description

Issue #8429 

Rewriting the SSH sample and its test using new libraries and hopefully avoiding the nasty test flakiness.

This commit contains the new sample and test for it. It only disables (by skipping) the test for old sample. The old sample is for now left as it is, because we first need to change the sample used on [this documentation page](https://cloud.google.com/compute/docs/tutorials/service-account-ssh).