fix: bring back region tag
process for changing the name for a region tag is to add the second one, update docs, then remove the old tag. So this requires both region tags currently. 

Second, remove the if statement. Even if this is not matching the intent of the other samples, this is the preferred method from @engelke