fix: add region tag, move code around for docs (#9952)
* fix: add region tag, move code around for docs

    pair with @nicain; this is not typical usage of if __name__ == .. but
    to make sure documentation and intent of the sample for atlas this
    should work.

Co-authored-by: Nick Cain <nicholascain@google.com>

* Fix: linting Update main.py

I deleted one too many empty lines.

* fix: change debug to False

* fix: remove old region tag that is no longer used

* fix: standardize on server

remote region tag server_setup

* fix: bring back region tag

process for changing the name for a region tag is to add the second one, update docs, then remove the old tag. So this requires both region tags currently. 

Second, remove the if statement. Even if this is not matching the intent of the other samples, this is the preferred method from @engelke

* fix: remove os as no longer using it

---------

Co-authored-by: Nick Cain <nicholascain@google.com>