feat: modernize Dataflow pipeline_with_dependencies (#12458)
* cleanup: structure dependecies using src/

Modernize example by providing a strucutre using src/
folder instead of flat folder structure. This is more
common on today's Python Best Practices.

See CL 663542630 for more details.

* feat: add dataflow example with pyproject.toml

The current pipeline_with_dependencies example is
using setup.py. A more up-to-date approach would
be using pyproject.toml. This folder contains a
copy of the original python_with_dependencies, but
leverages pyproject.toml for the package setup.

See CL 663542630 for additional context.

* fix: move _test files to root

For the tests to be recognized, they must be on the same
folder as the nox configurations.

* cleanup: remove pipeline_with_dependencies_toml

* feat: use toml file for packaging

* cleanup: regenerate requirements.txt

* cleanup: remove .egg-info

* fix: ignore egg-info

* nit: order files alphabetically

* docs: update README for requirements.txt

* review: address review feedback #12458

* cleanup: move main.py to root folder

* fix: fix path to main.py

* cleanup: add build-system to pyproject.toml