Update generate_self_signed_jwt.py (#12592)
* Update generate_self_signed_jwt.py

This is a bug fix in response to a listener report. The datetime object needs to return the current time in epoch

* fix(iap): refactor to use time and fix conversion

JWT claims require that 
IAT - "This is a timestamp formatted as whole seconds since January 01, 1970, which is a standard UNIX representation of time. The timestamp must be an integer (no decimals) and it must be in UTC."

This means using the time module is sufficient

---------

Co-authored-by: Jennifer Davis <sigje@google.com>